delete in details works functinally
