+++ components/ProductListItem/index.js
@@ -1,7 +1,6 @@
 import styled from "styled-components";
 import Link from "next/link";
-import { useState } from "react";
-import { StyledSubmitButton, StyledCancelButton } from "../Buttons";
+import DeleteConfirmation from "../DeleteConfirmation";
 
 const StyledListItem = styled.li`
   margin-block: 7px;
@@ -21,70 +20,12 @@ const StyledLink = styled(Link)`
     cursor: pointer;
   }
 `;
-const StyledDeleteIcon = styled.button`
-  width: 15%;
-  padding-block: 10px;
-  border: none;
-  background-color: transparent;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-const StyledDeleteConfirmation = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  column-gap: 5px;
-  padding: 5px;
-`;
-const StyledConfirmButton = styled(StyledSubmitButton)`
-  width: auto;
-  display: flex;
-  justify-content: center;
-  margin: 0px;
-  font-size: 14px;
-  padding: 5px 10px;
-  background-color: var(--deleteColor);
-  &:hover {
-    background-color: darkred;
-  }
-`;
-const StyledSmallCancelButton = styled(StyledCancelButton)`
-  width: auto;
-  display: flex;
-  justify-content: center;
-  margin: 0px;
-  font-size: 14px;
-  padding: 5px 10px;
-`;
 
 export default function ProductListItem({ product, onDeleteProduct }) {
-  const [showConfirmButtons, setShowConfirmButtons] = useState(false);
   return (
     <StyledListItem>
       <StyledLink href={`/products/${product._id}`}>{product.name} </StyledLink>
-      <StyledDeleteIcon
-        type="button"
-        onClick={() => setShowConfirmButtons(true)}
-        disabled={showConfirmButtons}
-      >
-        ✖️
-      </StyledDeleteIcon>
-      {showConfirmButtons && (
-        <StyledDeleteConfirmation>
-          <StyledSmallCancelButton
-            type="button"
-            onClick={() => setShowConfirmButtons(false)}
-          >
-            Cancel
-          </StyledSmallCancelButton>{" "}
-          <StyledConfirmButton
-            type="button"
-            onClick={() => onDeleteProduct(product._id)}
-          >
-            Confirm
-          </StyledConfirmButton>
-        </StyledDeleteConfirmation>
-      )}
+      <DeleteConfirmation product={product} onDeleteProduct={onDeleteProduct} />
     </StyledListItem>
   );
 }
